refactor: consolidate saml sp configs
- to be processed by a single class "SamlConfiguration" where
the @ConfigurationProperties(prefix="login.saml") annotation
has the ability to process all fields under the login.saml section
of UAA.yml
  - this is helpful because we can now centrally read, process,
  even validate all saml config fields under "login.saml"
  - pay attention to @ConfigurationProperties annotation's various
  requirements though: such as the private field names need to match
  the actually UAA.yml field name (e.g.: login.saml.fooBar -> private
  String fooBar); and that there need to be public setters and getters
  for each field
  - see: https://docs.spring.io/spring-boot/docs/current/reference/html/features.html#features.external-config.typesafe-configuration-properties.using-annotated-types
- the exception of the saml entity id, which in UAA.yml is somehow
outside of the login.saml context (set by login.entityID) so that
field stays under class SamlEntityIdConfiguration
